Gravel Pit Fix
Hopefully fixed issues with the gravel pit.
