+++ Assets/GravelPit/AlphaMat.mat.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 5e5cc9955d5c8864cb68713c4b5a75da
-timeCreated: 1473211787
+guid: 8e6e808b5a2d8b044bf92035549acfa1
+timeCreated: 1473331407
 licenseType: Free
 NativeFormatImporter:
   userData:
