+++ Assets/Resources/GravelPit/Gravel.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: fca9450eef46fce4c8c4954d0f38af28
-timeCreated: 1472805171
+guid: 6ddb58e5bc8f233478b015281ca69e15
+timeCreated: 1473337449
 licenseType: Free
 NativeFormatImporter:
   userData:
