--- Assets/Textures/Water.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: cf1144a14ef48fb448197badf3fb9eca
-folderAsset: yes
-timeCreated: 1472797514
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
